feat: restricted users CRUD

+++ backend/src/main/java/com/github/kzkaneoka/bbs/controller/UserController.java
@@ -1,12 +1,18 @@
 package com.github.kzkaneoka.bbs.controller;
 
+import com.github.kzkaneoka.bbs.enums.UserRole;
+import com.github.kzkaneoka.bbs.exception.RestControllerExceptionHandler;
 import com.github.kzkaneoka.bbs.model.User;
 import com.github.kzkaneoka.bbs.repository.UserRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -16,10 +22,13 @@ import java.util.UUID;
 @RequestMapping("/api")
 public class UserController {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
+
     @Autowired
     UserRepository userRepository;
 
     @GetMapping("/users")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<List<User>> getAllUsers() {
         List<User> users = new ArrayList<>();
         userRepository.findAll().forEach(users::add);
@@ -30,14 +39,26 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<User> getuserById(@PathVariable("id") UUID id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ResponseEntity<User> getUserById(@PathVariable("id") UUID id, Principal principal) {
         Optional<User> userData = userRepository.findById(id);
+        User loggedInUser = userRepository.findByUsername(principal.getName());
+        if (!loggedInUser.getRoles().stream().anyMatch(role -> role.getName().equals(UserRole.ROLE_ADMIN))
+                && !userData.get().getUsername().equals(principal.getName())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return new ResponseEntity<>(userData.get(), HttpStatus.OK);
     }
 
     @PutMapping("/users/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable("id") UUID id, @RequestBody User user) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ResponseEntity<User> updateUser(@PathVariable("id") UUID id, @RequestBody User user, Principal principal) {
         Optional<User> userData = userRepository.findById(id);
+        User loggedInUser = userRepository.findByUsername(principal.getName());
+        if (!loggedInUser.getRoles().stream().anyMatch(role -> role.getName().equals(UserRole.ROLE_ADMIN))
+                && !userData.get().getUsername().equals(principal.getName())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         User _user = userData.get();
         if (user.getUsername() != null && !user.getUsername().isEmpty()) {
             _user.setUsername(user.getUsername());
@@ -49,12 +70,20 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<HttpStatus> deleteUser(@PathVariable("id") UUID id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ResponseEntity<HttpStatus> deleteUser(@PathVariable("id") UUID id, Principal principal) {
+        Optional<User> userData = userRepository.findById(id);
+        User loggedInUser = userRepository.findByUsername(principal.getName());
+        if (!loggedInUser.getRoles().stream().anyMatch(role -> role.getName().equals(UserRole.ROLE_ADMIN))
+                && !userData.get().getUsername().equals(principal.getName())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         userRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @DeleteMapping("/users")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<HttpStatus> deleteAllUsers() {
         userRepository.deleteAll();
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
